requirements-test.txt
# Testing Dependencies for ScrollIntel Integration Tests

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0  # For parallel test execution
pytest-timeout>=2.1.0
pytest-json-report>=1.5.0

# Performance testing
pytest-benchmark>=4.0.0
psutil>=5.9.0

# Security testing
bandit>=1.7.5
safety>=2.3.0

# HTTP testing
httpx>=0.24.0
requests>=2.31.0

# Database testing
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0

# Mock and fixtures
responses>=0.23.0
factory-boy>=3.3.0
faker>=19.0.0

# Test data generation
pandas>=2.0.0
numpy>=1.24.0

# Container testing
testcontainers>=3.7.0
docker>=6.1.0

# Load testing
locust>=2.15.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.4.0

# Documentation testing
pytest-doctestplus>=1.0.0

# Reporting
pytest-html>=3.2.0
pytest-json>=0.4.0
allure-pytest>=2.13.0

# Environment management
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Async testing utilities
aioresponses>=0.7.4
asynctest>=0.13.0